+++ SpaceBattle/IUObject.cs
@@ -1,0 +1,7 @@
+﻿namespace SpaceBattle;
+
+public interface IUObject
+{
+    void SetProperty(string key, object value);
+    object GetProperty(string key);
+}
